requirements.txt
python-dotenv
dash
plotly
telethon
mysql-connector
pandas